Rust port L3 layout: TransistorList i32 → u16 (697KB → 350KB; +3.5%)
Sync the highest-RoI piece of the C# L3-layout series (commit ed3e637)
to the Rust port. Other items (group_buf, recalc_list shrink to u16,
NodeInfo SoA split) were verified marginal in C# — skipping them here.

Other size optimizations already in place in the Rust port from day one:
  recalc_hash[Next] is Vec<u8>  (C# took until ca49aa7 to match this)
  in_group is Vec<u8>           (C# took until e6824e5)

Storage: snapshot.rs loader reads i32 from the v4 wire format and casts
to u16 in memory. wire.rs / parallel.rs readers do `as i32` on read
(single µop on x64, free) so downstream code is unchanged.

Bench (full_palette.aprsnap, 500K hc, --fast-path, 5 runs):
  prior baseline (memory note rust-port-best-config):  best 48,622 hc/s
  with u16 TransistorList:                              best 50,306 hc/s
                                                        avg  50,061 hc/s
                                                        → +3.5% best

Bit-identical to C#: checksum 0x2682907736914E31 @ t=500000.

Rust port now crosses 50K hc/s on single-core. NES real-time still ~365×
away — confirms [[s4-route-single-instance]] wind-down conclusion.

+++ experiment/rust-poc/src/parallel.rs
@@ -73,7 +73,7 @@ impl ChipScratch {
 pub struct SharedView {
     pub node_states: *mut u8,
     pub node_infos: *const NodeInfo,
-    pub transistor_list: *const i32,
+    pub transistor_list: *const u16,
     pub flags_to_state: *const u8,
     pub chip_id: *const u8,
     pub target_to_handler: *const i32,
@@ -113,9 +113,9 @@ unsafe fn add_node_to_group_chip_aware(view: SharedView, scratch: &mut ChipScrat
     if ni.tlist_c1c2s != 0 {
         let mut p = ni.tlist_c1c2s as usize;
         loop {
-            let gate = *view.transistor_list.add(p);
+            let gate = *view.transistor_list.add(p) as i32;
             if gate == 0 { break; }
-            let other = *view.transistor_list.add(p + 1);
+            let other = *view.transistor_list.add(p + 1) as i32;
             p += 2;
             if *view.node_states.add(gate as usize) != 0 {
                 add_node_to_group_chip_aware(view, scratch, other);
@@ -126,7 +126,7 @@ unsafe fn add_node_to_group_chip_aware(view: SharedView, scratch: &mut ChipScrat
     if ni.tlist_c1gnd != 0 {
         let mut p = ni.tlist_c1gnd as usize;
         loop {
-            let gate = *view.transistor_list.add(p);
+            let gate = *view.transistor_list.add(p) as i32;
             if gate == 0 { break; }
             p += 1;
             if *view.node_states.add(gate as usize) != 0 { scratch.group_flags |= FLAG_GND; break; }
@@ -135,7 +135,7 @@ unsafe fn add_node_to_group_chip_aware(view: SharedView, scratch: &mut ChipScrat
     if ni.tlist_c1pwr != 0 {
         let mut p = ni.tlist_c1pwr as usize;
         loop {
-            let gate = *view.transistor_list.add(p);
+            let gate = *view.transistor_list.add(p) as i32;
             if gate == 0 { break; }
             p += 1;
             if *view.node_states.add(gate as usize) != 0 { scratch.group_flags |= FLAG_PWR; break; }
@@ -154,9 +154,9 @@ unsafe fn set_node_state(view: SharedView, scratch: &mut ChipScratch, nn: i32, n
     if ni.tlist_gates == 0 { return; }
     let mut p = ni.tlist_gates as usize;
     loop {
-        let c1 = *view.transistor_list.add(p);
+        let c1 = *view.transistor_list.add(p) as i32;
         if c1 == 0 { break; }
-        let c2 = *view.transistor_list.add(p + 1);
+        let c2 = *view.transistor_list.add(p + 1) as i32;
         p += 2;
         if c1 != view.npwr && c1 != view.ngnd {
             push_to_chip_bucket(view, scratch, c1);
